Working post and delete with curl and browser

+++ src/methods/post.cpp
@@ -1,4 +1,5 @@
 #include "httpResponse.hpp"
+#include <filesystem>
 #include "CGI.hpp"
 
 unordered_map<string, string> parse_form_data(const string &body) {
@@ -17,15 +18,18 @@ unordered_map<string, string> parse_form_data(const string &body) {
     return form_data;
 }
 
-bool fileExists(const string &path) {
-    struct stat buffer;
-    return (stat(path.c_str(), &buffer) == 0);
-}
 
 void printRequestHeaders(const map<string, string> &headers) {
     for (const auto &header : headers) {
         cout << header.first << ": " << header.second << endl;
     }
+}
+
+bool fileExists(string &path) {
+    if (filesystem::exists(path))
+        return true;
+    else
+        return false;
 }
 
 bool storageExist(const string& path) {
@@ -49,7 +53,6 @@ string removeBoundaries(const string& content, const string& boundary) {
 void writeFile(const string& storagePath, const string& fileName, const string& content, Request &request) {
     string fullPath = storagePath + fileName;
     
-    cout << "Full path: " << fullPath << endl;
     if (!filesystem::exists(storagePath))
         filesystem::create_directories(storagePath);
     if (filesystem::exists(fullPath)) {
@@ -77,8 +80,11 @@ bool findLastBoundary(const string& contentType, const string& boundary) {
 	return (true);
 }
 
-void execPythonScript(const std::string &file_path, const std::string &destination_folder) {
-    char *python_cgi = (char *)"/usr/bin/python3";
+void execScript(const std::string &file_path, const std::string &destination_folder) 
+{
+    // I think we are going to need to pass the request class as an argument of this function
+    char *python_cgi = (char *)"/usr/bin/python3"; 
+    // It makes more sense to me to put it in the html file
     char *script_path = (char *)"var/cgi-bin/upload_file_in_a_folder.cgi";  // Adjust path
 
     // Prepare the arguments (script name, file path, destination folder)
@@ -91,57 +97,83 @@ void execPythonScript(const std::string &file_path, const std::string &destinati
 }
 
 
-static void run_script(const string& script_path, const string& destination_folder) {
+static int run_script(const string& script_path, const string& destination_folder) {
+    /////////////////////////////////////////////////////////////
+    //  I think this function can accept the class reqest as an argument or maybe we can just send connection everywhere
     pid_t pid = fork();
     int	pipefd[2];
     int status;
 
     if (pipe(pipefd) == -1) {
         cerr << "Failed to create pipe" << endl;
-        return ;
+        return 1;
     }
     if (pid == -1) {
         cerr << "Failed to fork" << endl;
-        return ;
+        return 1;
     }
     if (pid == 0) {
-        cout << "Child process" << endl;
         close(pipefd[1]);
-        cout << "Script path: " << script_path << endl;
-        execPythonScript("Makefile" , destination_folder);
+        execScript(script_path , destination_folder);
     }
     else {
         cout << "Parent process" << endl;
         close(pipefd[0]);
         if (waitpid(pid, &status, 0) == -1) {
             cerr << "Failed to wait for child process" << endl;
-            return ;
+            return 1;
         }
         else {
             cout << "Child process exited with status " << status << endl;
+            // set error code maybe here /* Not Sure */
         }
     }
+    return 0;
 }
 
+string extract_file_name(const string &path) {
+    size_t posSlash = path.find_last_of('/');
+    string file = path.substr(posSlash + 1);
+    return file;
+}
 
-void post_method(int clientSocket, Request &request) {
+void post_method(int clientSocket, Request &request)
+{
     string uploadedFile;
     string storage = "www/storage/";
+    string fullPath;
 
-    cout << "In post method" << endl;
+    if (request.getPath() == "/www/deleteFile.html") // isCGI = true?
+	{
+        // switch to delete method
+        delete_method(clientSocket, request);
+        return ;
+    }
+    if (!request.getUploadedFile().empty())
+        fullPath = storage + request.getUploadedFile();
+    else
+        fullPath = storage + extract_file_name(request.getPath());
+    if (fileExists(fullPath)) {
+        request.setPath("www/fileExists.html");
+        return ;
+    }
+    // Once Jan's code is merged, we can use the body and go to the CGI script in both cases
+    /////////////////////////////////////////////////////////////////////////////////////////
     if (!request.getBoundary().empty() && request.getBytesCopied() <= stol(request.getHeaderValue("Content-Length"))){
         writeFile(storage, request.getUploadedFile(), request.getContentUploadFile(), request);
-        // execPythonScript(request.getUploadedFile(), request.getContentUploadFile());
-        // test_cgi("/home/ibehluli/Desktop/webserver/var/cgi-bin/upload_file_in_a_folder.cgi");
         request.setBytesCopied(request.getBytesCopied() + request.getMaxLengthUploadContent());
-        if (request.getBytesCopied() == stol(request.getHeaderValue("Content-Length"))) {
-            request.setPath("base.html");
-        }
+        if (request.getBytesCopied() == stol(request.getHeaderValue("Content-Length")))
+            request.setPath("www/fileUploaded.html");
+        else
+            request.setPath("www/fileNotUploaded.html");
     }
+    /////////////////////////////////////////////////////////////////////////////////////////
     else
     {
-        cout << "HELEOEOE" << endl;
-        run_script(request.getPath(), storage);
+        if (run_script(request.getPath(), storage))
+            request.setPath("www/fileNotUploaded.html");
+        else
+            request.setPath("www/fileUploaded.html");
         // cerr << RED << "Error with a post Request" << RESET << endl; 
     }
 }
